Take a another crack a fixing failing tests
Tests that check that a time is close to now +/1 1 sec sometimes fail.

I think this is because S3 always rounds the time down, and so if the time is close to the next second, it gets rounded down, and then if the time flips over to the next second before the test, the test fails. Usually the test time when it fails is just barely over the second mark.