Fix error when S3 object Etag is not an MD5
Objects saved by the blobstore should always have MD5s as Etags, but if data is
imported into S3 (say by the Minio mc client) that may not be true. When
getting or copying an S3 object, allow the Etag to not
be an MD5.

Storing an object still requires an MD5, since that's how we get the MD5 to put
into MongoDB. On a get or copy we can just use that MD5 instead of the S3 Etag.

Go formatted some test files for the first time also.